timeplot.ENH: add showToolTip property (moved from custom widget to RadTotDoseMonitor)
